Update package structure following simplifications
