--- src/aoai-simulated-api/src/aoai_simulated_api/record_replay/__init__.py
@@ -1,5 +1,0 @@
-# imports here allow aggregrating types under aoai_simulated_api.record_replay
-# pylint: disable=useless-import-alias
-from ._persistence import YamlRecordingPersister as YamlRecordingPersister
-from ._record_replay_handler import RecordReplayHandler as RecordReplayHandler
-from ._request_forwarder_config import get_default_forwarders as get_default_forwarders
