+++ src/aoai-simulated-api/src/aoai_simulated_api/record_replay/handler.py
@@ -8,14 +8,39 @@ import requests
 
 from aoai_simulated_api import constants
 from aoai_simulated_api.models import RequestContext
-
-from aoai_simulated_api.record_replay._hashing import get_request_hash, hash_request_parts
-from aoai_simulated_api.record_replay._models import RecordedResponse
-from aoai_simulated_api.record_replay._persistence import YamlRecordingPersister
+from aoai_simulated_api.record_replay.openai import forward_to_azure_openai
+from aoai_simulated_api.record_replay.models import RecordedResponse, get_request_hash, hash_request_parts
+from aoai_simulated_api.record_replay.persistence import YamlRecordingPersister
 
 logger = logging.getLogger(__name__)
 
 text_content_types = ["application/json", "application/text"]
+
+
+def get_default_forwarders() -> list[
+    Callable[
+        [RequestContext],
+        fastapi.Response
+        | Awaitable[fastapi.Response]
+        | requests.Response
+        | Awaitable[requests.Response]
+        | dict
+        | Awaitable[dict]
+        | None,
+    ]
+]:
+    # Return a list of functions to call when recording and no matching saved request is found
+    #
+    # If the function returns a Response object (from FastAPI or requests package)
+    # it will be used as the response for the request
+    #
+    # If the function returns a dict then it should have a "response" property
+    # with the response and a "persist" property that is True/False to indicate whether to persist the response
+    #
+    # If the function returns None, the next function in the list will be called
+    return [
+        forward_to_azure_openai,
+    ]
 
 
 class ForwardedResponse:
