+++ src/aoai-simulated-api/src/aoai_simulated_api/record_replay/models.py
@@ -1,4 +1,18 @@
+from dataclasses import dataclass
 from fastapi import Request
+
+
+@dataclass
+class RecordedResponse:
+    request_hash: int
+    status_code: int
+    status_message: str
+    headers: dict[str, list[str]]
+    body: str
+    duration_ms: int
+    # full_request currently here for compatibility with VCR serialization format
+    # it _is_ handy for human inspection to have the URL/body etc. in the recording
+    full_request: dict
 
 
 def hash_request_parts(method: str, url: str, body: bytes):
